+++ docs/source/changes.rst
@@ -2,6 +2,12 @@
 Change Log
 ==========
 
+?.?.?
+-----
+
+- Add Celery to the project to move non-critical-path event handling logic
+  to asynchronous tasks.
+
 3.4.2
 -----